Reverts code to v1.0.6
When I moved the code from SourceForge to GitHub, people naturally
started to use the trunk head to compile jp2a themselves.
Unfortunately, it contained a lot of work in progress code that didn't
work.

To fix that, I just just reverted to the last released SourceForge
version that worked. That should be commit 46a22..

It's been over ten years since I last had a look at this code. Oh boy,
how ugly it is! I've scanned through the changes for specific bug
fixes, but couldn't really find anything.

I just want to thank all the people who pointed out that the current
code didn't work, and a special thanks to all the maintainers over the
years (in particular, Eriberto!)! If you can please test if this
version works and let me know, I can mark it as a proper release, and
I can start to accept issues again and fix them.